+++ actions/registry.go
