+++ lib/history.go
@@ -4,9 +4,9 @@ import (
 	"fmt"
 	"net/rpc"
 
+	"github.com/qri-io/qri/actions"
 	"github.com/qri-io/qri/p2p"
 	"github.com/qri-io/qri/repo"
-	"github.com/qri-io/qri/repo/actions"
 )
 
 // HistoryRequests encapsulates business logic for the log
