+++ lib/search.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"net/rpc"
 
+	"github.com/qri-io/qri/actions"
 	"github.com/qri-io/qri/repo"
-	"github.com/qri-io/qri/repo/actions"
 	"github.com/qri-io/registry/regclient"
 )
